streamlit_app/requirements.txt
streamlit
torch
torchvision
Pillow
requests
streamlit-drawable-canvas
psycopg2-binary
